refactor: S10.10 hoist composite-cast in UTC offset formatter (MISRA 10.8)
Cast absoluteMinutes to uint32_t once at the top of
SolidSyslog_FormatNonZeroUtcOffset and use uint32_t arithmetic for the
hours/minutes split. Previous form cast each composite (absoluteMinutes
/ 60) and (absoluteMinutes % 60) — a different essential-type category
on a composite expression, which violates MISRA Rule 10.8.

cppcheck-misra: rule 10.8 findings 2 → 0.

+++ Core/Source/SolidSyslog.c
@@ -447,12 +447,12 @@ static inline void SolidSyslog_FormatUtcOffset(struct SolidSyslogFormatter* f, i
 
 static inline void SolidSyslog_FormatNonZeroUtcOffset(struct SolidSyslogFormatter* f, int16_t offsetMinutes)
 {
-    int16_t absoluteMinutes = SolidSyslog_AbsoluteInt16(offsetMinutes);
+    uint32_t absoluteMinutes = (uint32_t) SolidSyslog_AbsoluteInt16(offsetMinutes);
 
     SolidSyslogFormatter_AsciiCharacter(f, (offsetMinutes > 0) ? '+' : '-');
-    SolidSyslogFormatter_TwoDigit(f, (uint32_t) (absoluteMinutes / 60));
+    SolidSyslogFormatter_TwoDigit(f, absoluteMinutes / 60U);
     SolidSyslogFormatter_AsciiCharacter(f, ':');
-    SolidSyslogFormatter_TwoDigit(f, (uint32_t) (absoluteMinutes % 60));
+    SolidSyslogFormatter_TwoDigit(f, absoluteMinutes % 60U);
 }
 
 static inline int16_t SolidSyslog_AbsoluteInt16(int16_t value)
